backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
mysql-connector-python==8.1.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
scikit-learn==1.3.0
PyMySQL==1.1.0
SQLAlchemy==2.0.20
gunicorn==21.2.0
python-dotenv==1.0.0 